Replace GraalVM by the JAVET library

+++ src/main/java/org/orbisgis/demat/IRenderer.java
@@ -1,11 +1,9 @@
 package org.orbisgis.demat;
 
 
+import com.caoccao.javet.exceptions.JavetException;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import j2html.tags.DomContent;
-import org.graalvm.polyglot.Context;
-import org.graalvm.polyglot.Engine;
-import org.graalvm.polyglot.HostAccess;
-import org.graalvm.polyglot.Source;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -14,7 +12,6 @@ import java.io.IOException;
 import static j2html.TagCreator.*;
 
 public interface IRenderer {
-
 
     String getHTMLDirectory();
 
@@ -59,6 +56,10 @@ public interface IRenderer {
         return save(new File(path), true);
     }
 
+    default String save(File file) throws IOException {
+        return save(file, true);
+    }
+
     /**
      * Save into an html file
      *
@@ -67,31 +68,64 @@ public interface IRenderer {
      * @return
      */
     default String save(File outputFile, boolean delete) throws IOException {
-        String jsDirectory = getHTMLDirectory();
-        FileUtils.deployJSFiles(new File(jsDirectory));
-        if (outputFile == null) {
-            return null;
-        }
-        if (outputFile.exists()) {
-            if (delete) {
-                outputFile.delete();
-            } else {
-                return null;
+        if(FileUtils.isExtensionWellFormated(outputFile,"png")){
+            try {
+                return IOUtils.saveAsPNG(toJson(), getHTMLDirectory(), outputFile, delete);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
             }
         }
-        FileWriter fileWriter = new FileWriter(outputFile);
-        html(
-                head(
-                        meta().withCharset("UTF-8"),
-                        script().withSrc(jsDirectory+  File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[0]),
-                        script().withSrc(jsDirectory+  File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[1]),
-                        script().withSrc(jsDirectory+  File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[2])
-                ),
-                body(getDomElements())
-        ).render(fileWriter);
-        fileWriter.close();
-        return outputFile.getAbsolutePath();
+        else if(FileUtils.isExtensionWellFormated(outputFile,"svg")){
+            try {
+                return IOUtils.saveAsSVG(toJson(), getHTMLDirectory(), outputFile, delete);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            } catch (JavetException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        else if(FileUtils.isExtensionWellFormated(outputFile,"json")){
+                IOUtils.saveAsJSON(toJson(), outputFile, delete);
+        }
+        else if(FileUtils.isExtensionWellFormated(outputFile,"html")) {
+            String jsDirectory = getHTMLDirectory();
+            FileUtils.deployJSFiles(new File(jsDirectory));
+            if (outputFile == null) {
+                return null;
+            }
+            if (outputFile.exists()) {
+                if (delete) {
+                    outputFile.delete();
+                } else {
+                    return null;
+                }
+            }
+            FileWriter fileWriter = new FileWriter(outputFile);
+            html(
+                    head(
+                            meta().withCharset("UTF-8"),
+                            script().withSrc(jsDirectory + File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[0]),
+                            script().withSrc(jsDirectory + File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[1]),
+                            script().withSrc(jsDirectory + File.separator + FileUtils.JS_FOLDER + File.separator + FileUtils.JS_FILES[2])
+                    ),
+                    body(getDomElements())
+            ).render(fileWriter);
+            fileWriter.close();
+            return outputFile.getAbsolutePath();
+        }else {
+            throw new RuntimeException("Cannot save the file " + outputFile.getAbsolutePath());
+        }
+        return null;
     }
 
     DomContent getDomElements();
+
+    /**
+     * Build a json representation of the chart
+     *
+     * @return
+     * @throws JsonProcessingException
+     */
+    String toJson() throws JsonProcessingException;
+
 }
